ml-training/requirements.txt
lightgbm>=4.3
scikit-learn>=1.4
numpy>=1.26
pandas>=2.2
arch>=6.3
scipy>=1.13
psycopg2-binary>=2.9
skl2onnx>=1.16
onnxmltools>=1.12
onnxruntime>=1.18
matplotlib>=3.8
